macOS hangs when Accessibility Permission being revoked while running
