+++ LinearMouse/EventTap/EventTap.swift
@@ -96,6 +96,7 @@ extension EventTap {
             withExtendedLifetime(contextHolder) {
                 CGEvent.tapEnable(tap: tap, enable: false)
                 CFRunLoopRemoveSource(cfRunLoop, runLoopSource, .commonModes)
+                CFMachPortInvalidate(tap)
             }
         }
     }
